# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend REST API for golang nuryanto2121",
        "title": "Starter",
        "contact": {
            "name": "Nuryanto",
            "url": "https://www.linkedin.com/in/nuryanto-1b2721156/",
            "email": "nuryantofattih@gmail.com"
        },
        "license": {},
        "version": "1.0"
    },
    "paths": {
        "/capster/auth/change_password": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Change Password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "req param #changes are possible to adjust the form of the registration form from frontend",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.ResetPasswd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
        "/capster/auth/forgot": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Forgot Password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "req param #changes are possible to adjust the form of the registration form from frontend",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.ForgotForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
        "/capster/auth/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "req param #changes are possible to adjust the form of the registration form from frontend",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.LoginForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
        "/capster/auth/register": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "req param #changes are possible to adjust the form of the registration form from frontend",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.RegisterForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
        "/capster/auth/verify": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Verify",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "req param #changes are possible to adjust the form of the registration form from frontend",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.VerifyForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ForgotForm": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                }
            }
        },
        "models.LoginForm": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "pwd": {
                    "type": "string"
                }
            }
        },
...
        "models.RegisterForm": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "confirm_pwd": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pwd": {
                    "type": "string"
                },
                "user_type": {
                    "type": "string"
                }
            }
        },
        "models.ResetPasswd": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "confirm_pwd": {
                    "type": "string"
                },
                "pwd": {
                    "type": "string"
                }
            }
        },
...
        "models.VerifyForm": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "verify_code": {
                    "type": "string"
                }
            }
        },
        "tool.ResponseModel": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "description": "Code int         `json:\"code\"`",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}